fix: complete String->Long entity id migration across the actor layer
Finishes the id migration: closes the ~236 mechanical compile errors left
in the model/hybrid actor layer and micro simulation strategy, then
converts the last remaining String receive-boundary field
(ActorInteractionEvent.actorRefId, and its int64 wire counterpart in
communication.proto) so no business logic does .toLong/.toString on an
entity id anymore - only the two genuinely unavoidable Pekko-sharding
crossings do.

Also fixes two real correctness gaps surfaced along the way: reverts
NodeState.subwayStations/busStops back to String (they're route/line
labels, not ids, mistakenly swept into Long by an earlier pass), and
verifies the 0L "unset" sentinel is safe against actual id generation
(IdUtil/ActorSimulation.id never produces 0).

sbt compile and sbt test (280/280, 57 suites) both green.

+++ src/main/scala/core/actor/manager/DigitalTwinManager.scala
@@ -93,7 +93,7 @@ class DigitalTwinManager(timeManager: ActorRef)
           event = ActorInteractionEvent(
             tick          = 0L,
             lamportTick   = 0L,
-            actorRefId    = entityId,
+            actorRefId    = entityId.toLong,
             shardRefId    = entityId,
             actorPathRef  = self.path.name,
             actorClassType = getClass.getSimpleName,
